requirements.txt
Flask==2.1.1
psycopg2-binary==2.8.6
joblib==0.17.0
pandas==1.4.2
sklearn==0.0
openai==0.18.0